Add files via upload
Within the Copus(we)1copy.txt file one can find the full version of the analyzed corpus; while the corpusanalysis.py file contains all the code utilized for conducting the examination of the text in question. The visualisation.py has been similarly implemented into the analysis of the chosen text in order to create a visual depiction of the data acquired during the empirical analysis in the form of a figure, where two graphs are present, both concerning the utilization of 'we' pronouns within the examined corpus.